# docs/openapi/dalm-openapi.yaml
openapi: 3.1.0
info:
  title: DALM API
  version: 1.0.0
  summary: AI 사진 매칭 및 엽서 서비스 API
  description: |
    하루 한 장의 사진을 등록하면 AI가 최근 7일 내 다른 사용자의 사진에서
    닮은 순간을 찾고, 매칭된 사용자끼리 한 장의 엽서를 주고받는 서비스의 API입니다.

    ## 핵심 정책
    - 사용자당 하루 한 장의 활성 사진
    - 사진은 4:5 비율이며 등록 후 AI 적합성 검사를 수행
    - 통과한 사진은 최대 7일 동안 매칭 후보를 탐색
    - 한 사진은 최대 한 번만 매칭
    - 매칭당 각 사용자는 엽서를 최대 한 장만 발송
    - 정확한 위치, 연락처, 카카오 친구 목록은 수집하거나 공개하지 않음
servers:
  - url: http://localhost:8000/v1
    description: Local development
  - url: https://api.example.com/v1
    description: Production placeholder
...
paths:
  /auth/kakao:
    post:
      tags: [Auth]
      summary: 카카오 로그인
      operationId: loginWithKakao
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KakaoLoginRequest'
      responses:
        '200':
          description: 기존 회원 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '201':
          description: 신규 회원 인증 성공, 온보딩 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/AuthenticationFailed'
        '403':
          $ref: '#/components/responses/AccountRestricted'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Access Token 재발급
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 재발급 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/AuthenticationFailed'
  /auth/logout:
    post:
      tags: [Auth]
      summary: 로그아웃
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: 로그아웃 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: 인증이 필요하거나 토큰이 만료됨
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    AuthenticationFailed:
      description: 카카오 인증 또는 Refresh Token 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    AccountRestricted:
      description: 이용 제한 계정
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    KakaoLoginRequest:
      type: object
      required: [access_token]
      properties:
        access_token:
          type: string
          minLength: 1
          description: 카카오 SDK에서 발급받은 Access Token
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 1
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          description: Access Token 유효 시간(초)
    AuthUser:
      type: object
      required: [id, nickname, status]
      properties:
        id:
          type: string
          format: uuid
        nickname:
          oneOf:
            - type: string
            - type: 'null'
        status:
          type: string
          enum: [ACTIVE, RESTRICTED, WITHDRAWN]
    AuthData:
      type: object
      required: [is_new_user, onboarding_required, tokens, user]
      properties:
        is_new_user:
          type: boolean
        onboarding_required:
          type: boolean
        tokens:
          $ref: '#/components/schemas/TokenPair'
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AuthData'
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TokenPair'
...
    SuccessEnvelope:
      type: object
      required: [data, error]
      properties:
        data: {}
        error:
          type: 'null'
    ErrorDetail:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
          format: uuid
        fields:
          type: array
          items:
            type: object
            required: [field, reason]
            properties:
              field:
                type: string
              reason:
                type: string
    ErrorResponse:
      type: object
      required: [data, error]
      properties:
        data:
          type: 'null'
        error:
          $ref: '#/components/schemas/ErrorDetail'
security:
  - bearerAuth: []